requirements.txt
django==6.0.1
gunicorn
mysqlclient
django-cors-headers
whitenoise
djangorestframework